examples/flask/requirements.txt
pycloudfn==0.1.119
flask==0.12
google-cloud-bigquery==0.24.0
